cmd/snapd/tool/snapd-apparmor: move snapd-apparmor sources around
Move snapd-apparmor sources to a dedicated tool location.

Related: SNAPDENG-37091
